backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.14.2",
    "fastapi>=0.136.0",
    "langchain-groq>=1.1.2",
    "litellm>=1.83.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.44.0",
]
